refactor: correct file name typo
